Refactor LoadingView out to separate file; delay cursor entrance until previous animations complete
